LayoutTests/platform/win/accessibility/list-marker-role.html

<html>
    <head>
        <script src="../../../fast/js/resources/js-test-pre.js"></script>

        <script>
            if (window.testRunner)
                testRunner.dumpAsText();
        </script>
    </head>
    <body id="body">
        <p>
            This test verifies that, on Windows, list markers return the
            non-editable "text" role.
            <br><br>
            <a href="https://bugs.webkit.org/show_bug.cgi?id="></a>
        </p>

        <ul>
            <li id="listItem" tabindex="1">list item</li>
        </ul>

        <p id="console"></p>

        <script>
            if (window.accessibilityController) {
                document.getElementById("listItem").focus();

                var listMarker = accessibilityController.focusedElement.childAtIndex(0);
                shouldBe('listMarker.role', '"text"');
            }
        </script>
    </body>
</html>
